Add api to atobarai module
