+++ apps/np-atobarai/graphql/subscriptions/transaction-refund-requested.graphql
@@ -11,11 +11,21 @@ fragment TransactionRefundRequestedEvent on TransactionRefundRequested {
       channel {
         ...Channel
       }
+      totalPrice {
+        gross {
+          amount
+        }
+      }
     }
     order {
       id
       channel {
         ...Channel
+      }
+      total {
+        gross {
+          amount
+        }
       }
     }
   }
