+++ apps/np-atobarai/src/app/api/webhooks/saleor/fulfillment-tracking-number-updated/use-case-response.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 
-import { AtobaraiApiClientFulfillmentReportError } from "@/modules/atobarai/types";
+import { AtobaraiApiClientFulfillmentReportError } from "@/modules/atobarai/api/types";
 
 import { FulfillmentTrackingNumberUpdatedUseCaseResponse } from "./use-case-response";
