+++ apps/np-atobarai/src/app/api/webhooks/saleor/fulfillment-tracking-number-updated/use-case.test.ts
@@ -8,12 +8,12 @@ import { mockedAtobaraiTransactionId } from "@/__tests__/mocks/atobarai/mocked-a
 import { mockedSaleorApiUrl } from "@/__tests__/mocks/saleor/mocked-saleor-api-url";
 import { mockedSaleorAppId } from "@/__tests__/mocks/saleor/mocked-saleor-app-id";
 import { mockedFulfillmentTrackingNumberUpdatedEvent } from "@/__tests__/mocks/saleor-events/mocked-fulfillment-tracking-number-updated-event";
-import { createAtobaraiFulfillmentReportSuccessResponse } from "@/modules/atobarai/atobarai-fulfillment-report-success-response";
+import { createAtobaraiFulfillmentReportSuccessResponse } from "@/modules/atobarai/api/atobarai-fulfillment-report-success-response";
 import {
   AtobaraiApiClientFulfillmentReportError,
   IAtobaraiApiClient,
   IAtobaraiApiClientFactory,
-} from "@/modules/atobarai/types";
+} from "@/modules/atobarai/api/types";
 
 import { AppIsNotConfiguredResponse, MalformedRequestResponse } from "../saleor-webhook-responses";
 import { FulfillmentTrackingNumberUpdatedUseCase } from "./use-case";
