+++ apps/np-atobarai/src/app/api/webhooks/saleor/transaction-initialize-session/use-case-response.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 
 import { mockedAtobaraiTransactionId } from "@/__tests__/mocks/atobarai/mocked-atobarai-transaction-id";
-import { AtobaraiApiClientRegisterTransactionError } from "@/modules/atobarai/types";
+import { AtobaraiApiClientRegisterTransactionError } from "@/modules/atobarai/api/types";
 import {
   ChargeActionRequiredResult,
   ChargeFailureResult,
