+++ apps/np-atobarai/src/app/api/webhooks/saleor/transaction-process-session/use-case.test.ts
@@ -13,12 +13,12 @@ import {
   CreditCheckResult,
   FailedReason,
   PendingReason,
-} from "@/modules/atobarai/atobarai-transaction-success-response";
+} from "@/modules/atobarai/api/atobarai-transaction-success-response";
 import {
   AtobaraiApiClientChangeTransactionError,
   IAtobaraiApiClient,
   IAtobaraiApiClientFactory,
-} from "@/modules/atobarai/types";
+} from "@/modules/atobarai/api/types";
 import {
   ChargeActionRequiredResult,
   ChargeFailureResult,
