+++ apps/np-atobarai/src/modules/app-config/trpc-handlers/new-config-trpc-handler.ts
@@ -6,11 +6,11 @@ import { TRPCError } from "@trpc/server";
 import { RandomId } from "@/lib/random-id";
 import { AppChannelConfig } from "@/modules/app-config/app-config";
 import { newConfigInputSchema } from "@/modules/app-config/trpc-handlers/new-config-input-schema";
+import { IAtobaraiApiClientFactory } from "@/modules/atobarai/api/types";
 import { createAtobaraiMerchantCode } from "@/modules/atobarai/atobarai-merchant-code";
 import { createAtobaraiSecretSpCode } from "@/modules/atobarai/atobarai-secret-sp-code";
 import { createAtobaraiShippingCompanyCode } from "@/modules/atobarai/atobarai-shipping-company-code";
 import { createAtobaraiTerminalId } from "@/modules/atobarai/atobarai-terminal-id";
-import { IAtobaraiApiClientFactory } from "@/modules/atobarai/types";
 import { protectedClientProcedure } from "@/modules/trpc/protected-client-procedure";
 
 export class NewConfigTrpcHandler {
