+++ apps/np-atobarai/src/modules/atobarai/api/atobarai-change-transaction-payload.ts
@@ -1,20 +1,23 @@
 import { z } from "zod";
 
-import { SaleorTransactionToken } from "../saleor/saleor-transaction-token";
-import { AtobaraiCustomer, AtobaraiCustomerSchema } from "./atobarai-customer";
+import { SaleorTransactionToken } from "@/modules/saleor/saleor-transaction-token";
+
+import { AtobaraiCustomer, AtobaraiCustomerSchema } from "../atobarai-customer";
 import {
   AtobaraiDeliveryDestination,
   AtobaraiDeliveryDestinationSchema,
-} from "./atobarai-delivery-destination";
-import { AtobaraiGoods, AtobaraiGoodsSchema } from "./atobarai-goods";
-import { AtobaraiMoney } from "./atobarai-money";
-import { ATOBARAI_SETTLEMENT_TYPE } from "./atobarai-settelment-type";
-import { AtobaraiShopOrderDate } from "./atobarai-shop-order-date";
+} from "../atobarai-delivery-destination";
+import { AtobaraiGoods, AtobaraiGoodsSchema } from "../atobarai-goods";
+import { AtobaraiMoney } from "../atobarai-money";
+import { ATOBARAI_SETTLEMENT_TYPE } from "../atobarai-settelment-type";
+import { AtobaraiShopOrderDate } from "../atobarai-shop-order-date";
+import { AtobaraiTransactionId, AtobaraiTransactionIdSchema } from "../atobarai-transaction-id";
 
 const schema = z
   .object({
     transactions: z.array(
       z.object({
+        np_transaction_id: AtobaraiTransactionIdSchema,
         shop_transaction_id: z.string().max(40),
         shop_order_date: z.string().regex(/^\d{4}-\d{2}-\d{2}$/, {
           message: "Date must be in YYYY-MM-DD format",
@@ -27,19 +30,21 @@ const schema = z
       }),
     ),
   })
-  .brand("AtobaraiRegisterTransactionPayload");
+  .brand("AtobaraiChangeTransactionPayload");
 
-export const createAtobaraiRegisterTransactionPayload = (args: {
+export const createAtobaraiChangeTransactionPayload = (args: {
+  atobaraiTransactionId: AtobaraiTransactionId;
   saleorTransactionToken: SaleorTransactionToken;
   atobaraiMoney: AtobaraiMoney;
   atobaraiCustomer: AtobaraiCustomer;
   atobaraiDeliveryDestination: AtobaraiDeliveryDestination;
   atobaraiGoods: AtobaraiGoods;
   atobaraiShopOrderDate: AtobaraiShopOrderDate;
-}) =>
-  schema.parse({
+}): AtobaraiChangeTransactionPayload => {
+  return schema.parse({
     transactions: [
       {
+        np_transaction_id: args.atobaraiTransactionId,
         shop_transaction_id: args.saleorTransactionToken,
         shop_order_date: args.atobaraiShopOrderDate,
         settlement_type: ATOBARAI_SETTLEMENT_TYPE,
@@ -50,5 +55,6 @@ export const createAtobaraiRegisterTransactionPayload = (args: {
       },
     ],
   });
+};
 
-export type AtobaraiRegisterTransactionPayload = z.infer<typeof schema>;
+export type AtobaraiChangeTransactionPayload = z.infer<typeof schema>;
