+++ apps/np-atobarai/src/modules/atobarai/api/atobarai-fulfillment-report-success-response.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
 
-import { AtobaraiTransactionIdSchema } from "./atobarai-transaction-id";
+import { AtobaraiTransactionIdSchema } from "../atobarai-transaction-id";
 
 const schema = z
   .object({
