+++ apps/np-atobarai/src/modules/atobarai/api/atobarai-register-transaction-payload.test.ts
